Run download requirements script after we deploy the script to setup epirepo
because the download requirements script depend on some script on the epirepo scripts.

this is still an effort to fix the flakiness of this role. We also revert it
back these changes in the download-requirements.yaml playbook as this change
should be a shorted fix (not necessary a better fix)

If I have time I would re-work the whole this role, replacing large shell with
ansible for much better maintenance. And we would avoid this maintenance
nightmare.