derterministic-hosts-order-in-hostgroup
This is related to issue https://github.com/epiphany-platform/epiphany/issues/1702.

In short epiphany return the ansible hostgroups (for a features such as
postgres) with underterministic order between runs. Basically the first run for
example groups['postgres'][0] is mapped to vm-1 and the second run it may map
to vm-0.

This causes some roles (like postgres or kafka) which depends on the order to
select the primary node or identify node_id will fail if we re-run it.

This change attempt to solve it by sorting the result return by `get_ips_for_feature`.